pyproject.toml
[project]
name = "alpha4gate"
version = "0.1.0"
description = "StarCraft II Protoss bot with rule-based decisions and Claude strategic advisor"
requires-python = ">=3.12"
dependencies = [
    "burnysc2",
    "fastapi",
    "uvicorn",
    "websockets",
    "python-dotenv",
    "anthropic",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alpha4gate"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "sc2: tests that require a running StarCraft II instance",
    "mp: tests that require multiplayer lobby",
]
addopts = "-m 'not sc2'"

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
packages = ["alpha4gate"]
